fix ids seed db

+++ renter-api/Models/Models/Book.cs
@@ -15,10 +15,10 @@ namespace Models.Models
         [Required]
         [StringLength(13)]
         public string ISBN { get; set; }//Podobno teraz 13 znakw https://pl.wikipedia.org/wiki/International_Standard_Book_Number
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public virtual Author Author { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public virtual Category Category { get; set; }
 
         public string Description { get; set; }
