+++ renter-api/Models/Models/MovieRating.cs
@@ -8,7 +8,7 @@ namespace Models.Models
     public class MovieRating
     {
         public long Id { get; set; }
-        public int MovieId { get; set; }
+        public long MovieId { get; set; }
         public virtual Movie Movie { get; set; }
         public string UserId { get; set; }
         public virtual User User { get; set; }
